requirements.txt
aiogram>=3.0.0
SQLAlchemy>=2.0.0
APScheduler>=3.10.1
python-dotenv
pytest>=7.0.0
loguru>=0.7.0
pydantic>=2.0.0
PyYAML>=6.0
alembic>=1.13.0
arrow>=1.3.0
lark>=1.3.0